+++ synapse-core/src/main/java/de/otto/synapse/consumer/MessageDispatcher.java
@@ -105,6 +105,7 @@ public class MessageDispatcher implements Consumer<TextMessage> {
                         }
                     } catch (final Exception e) {
                         LOG.error(e.getMessage(), e);
+                        throw new IllegalStateException(e.getMessage(), e);
                     }
                 });
     }